+++ lib/src/conflicts.rs
@@ -50,8 +50,14 @@ use crate::merge::MergedTreeValue;
 use crate::repo_path::RepoPath;
 use crate::store::Store;
 
-/// Length of conflict markers.
-pub const CONFLICT_MARKER_LEN: usize = 7;
+/// Minimum length of conflict markers.
+pub const MIN_CONFLICT_MARKER_LEN: usize = 7;
+
+/// If a file already contains lines which look like conflict markers of length
+/// N, then the conflict markers we add will be of length (N + increment). This
+/// number is chosen to make the conflict markers noticeably longer than the
+/// existing markers.
+pub const CONFLICT_MARKER_LEN_INCREMENT: usize = 4;
 
 fn write_diff_hunks(hunks: &[DiffHunk], file: &mut dyn Write) -> io::Result<()> {
     for hunk in hunks {
@@ -259,15 +265,23 @@ impl ConflictMarkerKind {
     }
 }
 
+/// Represents a conflict marker parsed from the file. Conflict markers consist
+/// of a single ASCII character repeated for a certain length.
+struct ConflictMarker {
+    kind: ConflictMarkerKind,
+    len: usize,
+}
+
 /// Write a conflict marker to an output file.
 fn write_conflict_marker(
     output: &mut dyn Write,
     kind: ConflictMarkerKind,
+    len: usize,
     suffix_text: &str,
 ) -> io::Result<()> {
-    let conflict_marker: BString = iter::repeat(kind.to_byte())
-        .take(CONFLICT_MARKER_LEN)
-        .collect();
+    debug_assert!(len >= MIN_CONFLICT_MARKER_LEN);
+
+    let conflict_marker: BString = iter::repeat(kind.to_byte()).take(len).collect();
 
     if suffix_text.is_empty() {
         writeln!(output, "{conflict_marker}")
@@ -276,16 +290,12 @@ fn write_conflict_marker(
     }
 }
 
-/// Parse a conflict marker from a line of a file. The conflict marker must have
-/// the correct length (CONFLICT_MARKER_LEN).
-fn parse_conflict_marker(line: &[u8]) -> Option<ConflictMarkerKind> {
+/// Parse a conflict marker from a line of a file. The conflict marker may have
+/// any length (even less than MIN_CONFLICT_MARKER_LEN).
+fn parse_conflict_marker_any_len(line: &[u8]) -> Option<ConflictMarker> {
     let first_byte = *line.first()?;
     let kind = ConflictMarkerKind::parse_byte(first_byte)?;
     let len = line.iter().take_while(|&&b| b == first_byte).count();
-
-    if len != CONFLICT_MARKER_LEN {
-        return None;
-    }
 
     if let Some(next_byte) = line.get(len) {
         // If there is a character after the marker, it must be ASCII whitespace
@@ -294,19 +304,66 @@ fn parse_conflict_marker(line: &[u8]) -> Option<ConflictMarkerKind> {
         }
     }
 
-    Some(kind)
+    Some(ConflictMarker { kind, len })
 }
 
+/// Parse a conflict marker, expecting it to have a certain length. Any conflict
+/// markers which have a different length are ignored.
+fn parse_conflict_marker(line: &[u8], expected_len: usize) -> Option<ConflictMarkerKind> {
+    debug_assert!(expected_len >= MIN_CONFLICT_MARKER_LEN);
+
+    parse_conflict_marker_any_len(line)
+        .filter(|marker| marker.len == expected_len)
+        .map(|marker| marker.kind)
+}
+
+/// Find the longest conflict marker in a file.
+pub fn find_max_conflict_marker_len(file: &[u8]) -> usize {
+    file.lines_with_terminator()
+        .filter_map(parse_conflict_marker_any_len)
+        .map(|marker| marker.len)
+        .max()
+        .unwrap_or_default()
+}
+
+/// Choose an appropriate conflict marker length to use for a file. The conflict
+/// marker length will be at least MIN_CONFLICT_MARKER_LENGTH, and is guaranteed
+/// to be longer than any existing markers in the file.
+pub fn choose_conflict_marker_len<T: AsRef<[u8]>>(single_hunk: &Merge<T>) -> usize {
+    let max_existing_marker_len = single_hunk
+        .iter()
+        .map(|term| find_max_conflict_marker_len(term.as_ref()))
+        .max()
+        .unwrap_or_default();
+
+    if max_existing_marker_len < MIN_CONFLICT_MARKER_LEN {
+        // Use the normal conflict marker length if it's unambiguous.
+        MIN_CONFLICT_MARKER_LEN
+    } else {
+        // If there are existing markers in the file, we want to make our conflict
+        // markers longer than them. We add CONFLICT_MARKER_LEN_INCREMENT to ensure
+        // they're longer by a visually noticeable amount.
+        max_existing_marker_len.saturating_add(CONFLICT_MARKER_LEN_INCREMENT)
+    }
+}
+
+// Returns conflict marker length used for materializing conflicts, if there
+// were any conflicts to materialize.
 pub fn materialize_merge_result<T: AsRef<[u8]>>(
     single_hunk: &Merge<T>,
     conflict_marker_style: ConflictMarkerStyle,
     output: &mut dyn Write,
-) -> io::Result<()> {
+) -> io::Result<Option<usize>> {
     let merge_result = files::merge(single_hunk);
     match &merge_result {
-        MergeResult::Resolved(content) => output.write_all(content),
+        MergeResult::Resolved(content) => {
+            output.write_all(content)?;
+            Ok(None)
+        }
         MergeResult::Conflict(hunks) => {
-            materialize_conflict_hunks(hunks, conflict_marker_style, output)
+            let conflict_marker_len = choose_conflict_marker_len(single_hunk);
+            materialize_conflict_hunks(hunks, conflict_marker_style, conflict_marker_len, output)?;
+            Ok(Some(conflict_marker_len))
         }
     }
 }
@@ -319,9 +376,15 @@ pub fn materialize_merge_result_to_bytes<T: AsRef<[u8]>>(
     match merge_result {
         MergeResult::Resolved(content) => content,
         MergeResult::Conflict(hunks) => {
+            let conflict_marker_len = choose_conflict_marker_len(single_hunk);
             let mut output = Vec::new();
-            materialize_conflict_hunks(&hunks, conflict_marker_style, &mut output)
-                .expect("writing to an in-memory buffer should never fail");
+            materialize_conflict_hunks(
+                &hunks,
+                conflict_marker_style,
+                conflict_marker_len,
+                &mut output,
+            )
+            .expect("writing to an in-memory buffer should never fail");
             output.into()
         }
     }
@@ -330,6 +393,7 @@ pub fn materialize_merge_result_to_bytes<T: AsRef<[u8]>>(
 fn materialize_conflict_hunks(
     hunks: &[Merge<BString>],
     conflict_marker_style: ConflictMarkerStyle,
+    conflict_marker_len: usize,
     output: &mut dyn Write,
 ) -> io::Result<()> {
     let num_conflicts = hunks
@@ -347,13 +411,21 @@ fn materialize_conflict_hunks(
             match (conflict_marker_style, hunk.as_slice()) {
                 // 2-sided conflicts can use Git-style conflict markers
                 (ConflictMarkerStyle::Git, [left, base, right]) => {
-                    materialize_git_style_conflict(left, base, right, &conflict_info, output)?;
+                    materialize_git_style_conflict(
+                        left,
+                        base,
+                        right,
+                        &conflict_info,
+                        conflict_marker_len,
+                        output,
+                    )?;
                 }
                 _ => {
                     materialize_jj_style_conflict(
                         hunk,
                         &conflict_info,
                         conflict_marker_style,
+                        conflict_marker_len,
                         output,
                     )?;
                 }
@@ -368,22 +440,35 @@ fn materialize_git_style_conflict(
     base: &[u8],
     right: &[u8],
     conflict_info: &str,
+    conflict_marker_len: usize,
     output: &mut dyn Write,
 ) -> io::Result<()> {
     write_conflict_marker(
         output,
         ConflictMarkerKind::ConflictStart,
+        conflict_marker_len,
         &format!("Side #1 ({conflict_info})"),
     )?;
     output.write_all(left)?;
-    write_conflict_marker(output, ConflictMarkerKind::GitAncestor, "Base")?;
+    write_conflict_marker(
+        output,
+        ConflictMarkerKind::GitAncestor,
+        conflict_marker_len,
+        "Base",
+    )?;
     output.write_all(base)?;
     // VS Code doesn't seem to support any trailing text on the separator line
-    write_conflict_marker(output, ConflictMarkerKind::GitSeparator, "")?;
+    write_conflict_marker(
+        output,
+        ConflictMarkerKind::GitSeparator,
+        conflict_marker_len,
+        "",
+    )?;
     output.write_all(right)?;
     write_conflict_marker(
         output,
         ConflictMarkerKind::ConflictEnd,
+        conflict_marker_len,
         &format!("Side #2 ({conflict_info} ends)"),
     )?;
 
@@ -394,44 +479,49 @@ fn materialize_jj_style_conflict(
     hunk: &Merge<BString>,
     conflict_info: &str,
     conflict_marker_style: ConflictMarkerStyle,
+    conflict_marker_len: usize,
     output: &mut dyn Write,
 ) -> io::Result<()> {
     // Write a positive snapshot (side) of a conflict
-    fn write_side(add_index: usize, data: &[u8], output: &mut dyn Write) -> io::Result<()> {
+    let write_side = |add_index: usize, data: &[u8], output: &mut dyn Write| {
         write_conflict_marker(
             output,
             ConflictMarkerKind::Add,
+            conflict_marker_len,
             &format!("Contents of side #{}", add_index + 1),
         )?;
         output.write_all(data)
-    }
+    };
 
     // Write a negative snapshot (base) of a conflict
-    fn write_base(base_str: &str, data: &[u8], output: &mut dyn Write) -> io::Result<()> {
+    let write_base = |base_str: &str, data: &[u8], output: &mut dyn Write| {
         write_conflict_marker(
             output,
             ConflictMarkerKind::Remove,
+            conflict_marker_len,
             &format!("Contents of {base_str}"),
         )?;
         output.write_all(data)
-    }
+    };
 
     // Write a diff from a negative term to a positive term
-    fn write_diff(
-        base_str: &str,
-        add_index: usize,
-        diff: &[DiffHunk],
-        output: &mut dyn Write,
-    ) -> io::Result<()> {
-        write_conflict_marker(
-            output,
-            ConflictMarkerKind::Diff,
-            &format!("Changes from {base_str} to side #{}", add_index + 1),
-        )?;
-        write_diff_hunks(diff, output)
-    }
+    let write_diff =
+        |base_str: &str, add_index: usize, diff: &[DiffHunk], output: &mut dyn Write| {
+            write_conflict_marker(
+                output,
+                ConflictMarkerKind::Diff,
+                conflict_marker_len,
+                &format!("Changes from {base_str} to side #{}", add_index + 1),
+            )?;
+            write_diff_hunks(diff, output)
+        };
 
-    write_conflict_marker(output, ConflictMarkerKind::ConflictStart, conflict_info)?;
+    write_conflict_marker(
+        output,
+        ConflictMarkerKind::ConflictStart,
+        conflict_marker_len,
+        conflict_info,
+    )?;
     let mut add_index = 0;
     for (base_index, left) in hunk.removes().enumerate() {
         // The vast majority of conflicts one actually tries to resolve manually have 1
@@ -484,6 +574,7 @@ fn materialize_jj_style_conflict(
     write_conflict_marker(
         output,
         ConflictMarkerKind::ConflictEnd,
+        conflict_marker_len,
         &format!("{conflict_info} ends"),
     )?;
     Ok(())
@@ -532,9 +623,16 @@ pub fn materialized_diff_stream<'a>(
 /// has to provide the expected number of merge sides (adds). Conflict
 /// markers that are otherwise valid will be considered invalid if
 /// they don't have the expected arity.
+///
+/// All conflict markers in the file must have the expected length. Any conflict
+/// markers with a different length will be ignored.
 // TODO: "parse" is not usually the opposite of "materialize", so maybe we
 // should rename them to "serialize" and "deserialize"?
-pub fn parse_conflict(input: &[u8], num_sides: usize) -> Option<Vec<Merge<BString>>> {
+pub fn parse_conflict(
+    input: &[u8],
+    num_sides: usize,
+    conflict_marker_len: usize,
+) -> Option<Vec<Merge<BString>>> {
     if input.is_empty() {
         return None;
     }
@@ -544,7 +642,7 @@ pub fn parse_conflict(input: &[u8], num_sides: usize) -> Option<Vec<Merge<BStrin
     let mut conflict_start = None;
     let mut conflict_start_len = 0;
     for line in input.lines_with_terminator() {
-        match parse_conflict_marker(line) {
+        match parse_conflict_marker(line, conflict_marker_len) {
             Some(ConflictMarkerKind::ConflictStart) => {
                 conflict_start = Some(pos);
                 conflict_start_len = line.len();
@@ -552,7 +650,7 @@ pub fn parse_conflict(input: &[u8], num_sides: usize) -> Option<Vec<Merge<BStrin
             Some(ConflictMarkerKind::ConflictEnd) => {
                 if let Some(conflict_start_index) = conflict_start {
                     let conflict_body = &input[conflict_start_index + conflict_start_len..pos];
-                    let hunk = parse_conflict_hunk(conflict_body);
+                    let hunk = parse_conflict_hunk(conflict_body, conflict_marker_len);
                     if hunk.num_sides() == num_sides {
                         let resolved_slice = &input[resolved_start..conflict_start_index];
                         if !resolved_slice.is_empty() {
@@ -584,27 +682,29 @@ pub fn parse_conflict(input: &[u8], num_sides: usize) -> Option<Vec<Merge<BStrin
 /// with other conflict marker styles from being parsed successfully. The
 /// conflict marker style to use for parsing is determined based on the first
 /// line of the hunk.
-fn parse_conflict_hunk(input: &[u8]) -> Merge<BString> {
+fn parse_conflict_hunk(input: &[u8], conflict_marker_len: usize) -> Merge<BString> {
     // If the hunk starts with a conflict marker, find its first character
     let initial_conflict_marker = input
         .lines_with_terminator()
         .next()
-        .and_then(parse_conflict_marker);
+        .and_then(|line| parse_conflict_marker(line, conflict_marker_len));
 
     match initial_conflict_marker {
         // JJ-style conflicts must start with one of these 3 conflict marker lines
         Some(ConflictMarkerKind::Diff | ConflictMarkerKind::Remove | ConflictMarkerKind::Add) => {
-            parse_jj_style_conflict_hunk(input)
+            parse_jj_style_conflict_hunk(input, conflict_marker_len)
         }
         // Git-style conflicts either must not start with a conflict marker line, or must start with
         // the "|||||||" conflict marker line (if the first side was empty)
-        None | Some(ConflictMarkerKind::GitAncestor) => parse_git_style_conflict_hunk(input),
+        None | Some(ConflictMarkerKind::GitAncestor) => {
+            parse_git_style_conflict_hunk(input, conflict_marker_len)
+        }
         // No other conflict markers are allowed at the start of a hunk
         Some(_) => Merge::resolved(BString::new(vec![])),
     }
 }
 
-fn parse_jj_style_conflict_hunk(input: &[u8]) -> Merge<BString> {
+fn parse_jj_style_conflict_hunk(input: &[u8], conflict_marker_len: usize) -> Merge<BString> {
     enum State {
         Diff,
         Remove,
@@ -615,7 +715,7 @@ fn parse_jj_style_conflict_hunk(input: &[u8]) -> Merge<BString> {
     let mut removes = vec![];
     let mut adds = vec![];
     for line in input.lines_with_terminator() {
-        match parse_conflict_marker(line) {
+        match parse_conflict_marker(line, conflict_marker_len) {
             Some(ConflictMarkerKind::Diff) => {
                 state = State::Diff;
                 removes.push(BString::new(vec![]));
@@ -675,7 +775,7 @@ fn parse_jj_style_conflict_hunk(input: &[u8]) -> Merge<BString> {
     }
 }
 
-fn parse_git_style_conflict_hunk(input: &[u8]) -> Merge<BString> {
+fn parse_git_style_conflict_hunk(input: &[u8], conflict_marker_len: usize) -> Merge<BString> {
     #[derive(PartialEq, Eq)]
     enum State {
         Left,
@@ -687,7 +787,7 @@ fn parse_git_style_conflict_hunk(input: &[u8]) -> Merge<BString> {
     let mut base = BString::new(vec![]);
     let mut right = BString::new(vec![]);
     for line in input.lines_with_terminator() {
-        match parse_conflict_marker(line) {
+        match parse_conflict_marker(line, conflict_marker_len) {
             Some(ConflictMarkerKind::GitAncestor) => {
                 if state == State::Left {
                     state = State::Base;
@@ -743,21 +843,46 @@ pub async fn update_from_content(
     // copy.
     let mut old_content = Vec::with_capacity(content.len());
     let merge_hunk = extract_as_single_hunk(simplified_file_ids, store, path).await?;
-    materialize_merge_result(&merge_hunk, conflict_marker_style, &mut old_content).unwrap();
+    let materialized_marker_len =
+        materialize_merge_result(&merge_hunk, conflict_marker_style, &mut old_content).unwrap();
     if content == old_content {
         return Ok(file_ids.clone());
     }
+
+    let max_content_marker_len = find_max_conflict_marker_len(content);
+
+    // If all conflict markers in the file content are shorter than the materialized
+    // conflict markers, then the file is considered resolved.
+    if max_content_marker_len < materialized_marker_len.unwrap_or(MIN_CONFLICT_MARKER_LEN) {
+        let file_id = store.write_file(path, &mut &content[..]).await?;
+        return Ok(Merge::normal(file_id));
+    }
+
+    // Only accept the longest conflict markers in the file, also requiring that
+    // they are at least as long as the materialized conflict markers. We need to
+    // allow conflict markers to be longer than the materialized conflict markers,
+    // since the user could have removed the "fake" conflict markers from the file
+    // while editing, in which case their existing conflict markers will be
+    // considered longer than necessary the next time the file is snapshotted.
+    let expected_marker_len = materialized_marker_len
+        .unwrap_or_else(|| choose_conflict_marker_len(&merge_hunk))
+        .max(max_content_marker_len);
 
     // Parse conflicts from the new content using the arity of the simplified
     // conflicts initially. If unsuccessful, attempt to parse conflicts from with
     // the arity of the unsimplified conflicts since such a conflict may be
     // present in the working copy if written by an earlier version of jj.
     let (used_file_ids, hunks) = 'hunks: {
-        if let Some(hunks) = parse_conflict(content, simplified_file_ids.num_sides()) {
+        if let Some(hunks) = parse_conflict(
+            content,
+            simplified_file_ids.num_sides(),
+            expected_marker_len,
+        ) {
             break 'hunks (simplified_file_ids, hunks);
         };
         if simplified_file_ids.num_sides() != file_ids.num_sides() {
-            if let Some(hunks) = parse_conflict(content, file_ids.num_sides()) {
+            if let Some(hunks) = parse_conflict(content, file_ids.num_sides(), expected_marker_len)
+            {
                 break 'hunks (file_ids, hunks);
             };
         };